+++ kitty/options/parse.py
@@ -8,7 +8,7 @@ from kitty.conf.utils import (
 )
 from kitty.options.utils import (
     active_tab_title_template, adjust_baseline, adjust_line_height, allow_hyperlinks,
-    allow_remote_control, box_drawing_scale, clear_all_mouse_shortcuts, clear_all_shortcuts,
+    allow_remote_control, box_drawing_scale, clear_all_mouse_actions, clear_all_shortcuts,
     clipboard_control, config_or_absolute_path, copy_on_select, cursor_text_color,
     deprecated_hide_window_decorations_aliases, deprecated_macos_show_window_title_in_menubar_alias,
     deprecated_send_text, disable_ligatures, edge_width, env, font_features, hide_window_decorations,
@@ -90,11 +90,11 @@ class Parser:
     def box_drawing_scale(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['box_drawing_scale'] = box_drawing_scale(val)
 
+    def clear_all_mouse_actions(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
+        clear_all_mouse_actions(val, ans)
+
     def clear_all_shortcuts(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         clear_all_shortcuts(val, ans)
-
-    def clear_all_mouse_shortcuts(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
-        clear_all_mouse_shortcuts(val, ans)
 
     def click_interval(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['click_interval'] = float(val)